+++ spec/support/test_helpers.rb
@@ -45,4 +45,17 @@ module TestHelpers
   ensure
     ENV['ENV'] = @original_env
   end
+
+  def scheme_date_for(text)
+    case text.downcase.strip
+      when 'scheme 11' then
+        Settings.agfs_scheme_11_release_date.strftime
+      when 'scheme 10' || 'post agfs reform' then
+        Settings.agfs_fee_reform_release_date.strftime
+      when 'lgfs' then
+        "2016-04-01"
+      else
+        "2016-01-01"
+    end
+  end
 end
